wvtest.py: when exceptions are thrown, choose a better stack dump.
We already print the entire stack trace - the useful part - but we were
printing a kind of useless line number in the wvtest result line: the line
from which the exception was thrown.  That's kind of weird, and doesn't help
indicate which test threw the exception.  So let's pull out the current line
number from the running @wvtest function instead.

That's not perfect either, in case the function called a sub-function that
produced the exception, but it's somewhat more useful at least.  You should
still read the full exception printout for maximum accuracy.