Update wvtest.py to the new version that does os.chdir().
It was too annoying to test submodules, because if any of them opened a
file, they always had to mention the full path from the wvtest.py (which was
usually at the root of the whole project) to the data file (which was
usually in the t/ subdir of the submodule).  So if you extracted a submodule
into its own project, all the paths would be wrong.

Now wvtest.py does os.chdir() into the t/ subdir being tested.  If you don't
like that, you can always chdir into a parent directory.