xfail/xpass/skip
Sometimes it is not possible to get everything in shape, and for such
situations testing tools usually have "fail, but we know that" result.

Dejagnu has XFAIL, and in Git they have test_expect_failure which is
printed as "still have N known breakage(s)"), etc...

Your approach to minimalistic testing looks pretty, but in my case I have
some tests which do not pass on win32, and I'd like not to concentrate
on them for the time being -> so something like Dejagnu XFAIL is needed.

I implemented one, but I have to say I don't know perl, and I had to
play with colors (we now have to highligh passing tests that were
expected to fail, to say look: it works now), etc... And
documentation is updated, but your style is yours, so please feel free
to reword it.

Please have a look. If we could include the idea into wvtest, that would
be useful.

Thanks,
Kirill