Add wvtest for C
For some C programs it is not possible to use C++ implementation
because g++ compiler doesn't support "Designated Initializers"
extension of GCC. If this extension is used in C code, the code cannot
be compiled by g++ and therefore it cannot contain WVPASS and similar
macros implemented in C++.

This patch adds C support which is derived the C++ one. The code
relies on "variable attributes" (another GCC extension) to run all the
functions defined with WVTEST_MAIN(). C++ API uses constructors of
global variables to find all these functions but in C, something else
must be used.