Fix the C API to work under OS X
Using variable attributes to put data structures to a separate section
doesn't work as expected under OS X. The code doesn't even compile.

This patch removes the use of variable attributes and uses instead
function attribute "constructor". The functions marked by this attribute
are called before main() similarly as C++ constructors of global
variables.

These functions are used to register all the tests defined with
WVTEST_MAIN so that we can run all of them.

Thanks to Jimmy Tang for helping with the fix.