C: Add simple tests for double numbers
The following tests are an outright copy and pastes of WVPASSEQ()
and WVPASSNE() and could probably be made smarter.

WVPASSEQ_DOUBLE(a, b, c) - where 'a' and 'b' are two numbers that
you wish to check if they are equal within a tolerance of 'c'

WVPASSNE_DOUBLE(a, b, c) - where 'a' and 'b' are two numbers that
you wish to check if they are not equal within a tolerance of 'c'

Note that 'c' tolerance should always be positive otherwise it is
confusing.