python: Implement WVXFAIL and @wvskip
Add abilities to skip and mark test with known breakage to python
driver.

Since there are no macros in python, skipping can be only implemented as
function decorator, not WVSKIP(code) - that code would execute then.

The driver implements results counting, and that is extended for couning
xfail/xpass/skip as well, although I don't get the rationale behind this
- results counting is, and should be, done in wvtestrun!