dotnet: Implement WVXFAIL and WVSKIP
Add abilities to skip and mark test with known breakage to dotnet
driver.

The driver implements results counting, and that is extended for counting
xfail/xpass/skip as well, although I don't get the rationale behind this
results counting is, and should be, done in wvtestrun!

Interaction with expect_fail is marked as TODO because rationale behind
it is not well understood - if it expected to fail - that's WVFAIL, but
there is no WVNOEXCEPT, and that expect_next_failure prefix is used. I
decided to better do it simple first, and then, if needed, it could be
adjusted (expect_next_failure() is unique to dotnet driver).

This is my first experience with C#, so please don't beat me too much.