c: Implement WVXFAIL and WVSKIP
Port changes to add abilities to skip and mark test with known breakage
from C++ driver to C driver.

We did not need to refactor print_result to accept string result for C -
it's already there in print_result_str().

The driver implements results counting, and that is extended for counting
xfail/xpass/skip as well, although I don't get the rationale behind this
- results counting is, and should be, done in wvtestrun!